requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
backoff
bcrypt==4.0.1
beautifulsoup4
blosc2
build
cachetools
Cerberus
certifi
cffi
cftime
chardet
charset-normalizer
click
colorama
coloredlogs
contourpy
cryptography
cycler
dataclasses-json
distro
durationpy
ecdsa
et_xmlfile
faiss-cpu
fastapi
filelock
flatbuffers
fonttools
frozenlist
fsspec
google-auth
googleapis-common-protos
greenlet
grpcio
h11
h5py
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
imageio
importlib_metadata
importlib_resources
Jinja2
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kiwisolver
kubernetes
langchain
langchain-classic
langchain-community
langchain-core
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lxml
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
mmh3
mpmath
msgpack
multidict
mypy_extensions
narwhals
ndindex
netCDF4
networkx
nltk
numexpr
numpy
oauthlib
onnxruntime
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-proto
opentelemetry-semantic-conventions
orjson
ormsgpack
overrides
packaging
pandas
passlib==1.7.4
patsy
pdfminer.six
pdfplumber
pillow
platformdirs
plotly
posthog
propcache
protobuf
py-cpuinfo
pyarrow
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
pyparsing
pypdf
PyPDF2
PyMuPDF  # Fast PDF parsing (5-10x faster than PyPDF2) - required for optimized RAG uploads
pypdfium2
PyPika
pyproject_hooks
pyreadline3
python-dateutil
python-docx
python-dotenv
python-pptx
python-jose
python-multipart
pytz
PyYAML
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
seaborn
sentence-transformers
rank-bm25  # For hybrid search (keyword + semantic)
setuptools
shellingham
six
sniffio
soupsieve
SQLAlchemy
sse-starlette
starlette
statsmodels
sympy
tables
tavily-python
tenacity
textblob
threadpoolctl
tiktoken
tokenizers
torch
tqdm
transformers
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
watchfiles
websocket-client
websockets
xarray
xlrd
xlwt
xxhash
yarl
zipp
zstandard
llama-cpp-python
pip-system-certs
aider-chat
